some changes to edit
