+++ src/components/admin/EditCreatedGame.js
@@ -6,35 +6,15 @@ import { yupResolver } from "@hookform/resolvers/yup";
 import { createHeaders } from "./CreateHeaders";
 import FormMessage from "../../common/FormMessage";
 import { useAuth0 } from "@auth0/auth0-react";
+import { useParams } from "react-router-dom";
 
-const apiUrl = "https://hvz-api-noroff.herokuapp.com/game";
+const apiUrl = "https://hvz-api-noroff.herokuapp.com/api/v1/games";
 
 const schema = yup.object().shape({
   gameTitle: yup
     .string()
     .required("Title is required")
     .min(3, "Title must be at least 3 characters"),
-  nw_lat: yup
-    .number()
-    .integer("Value must be an integer")
-    .required("Please enter a northwest latitude"),
-  nw_lng: yup
-    .number()
-    .integer("Value must be an integer")
-    .required("Please enter a northwest longitude"),
-  se_lat: yup
-    .number()
-    .integer("Value must be an integer")
-    .required("Please enter a southeast latitude"),
-  se_lng: yup
-    .number()
-    .integer("Value must be an integer")
-    .required("Please enter a southeast longitude"),
-  gameDescription: yup
-    .string()
-    .required("Please enter a description")
-    .min(20, "Dame description must be at least 20 characters long")
-    .max(200, "Game description must be at most 200 characters long"),
 });
 
 function EditCreatedGame() {
@@ -43,6 +23,7 @@ function EditCreatedGame() {
   const [postError, setPostError] = useState(null);
   const [postSuccess, setPostSuccess] = useState(false);
   const { getAccessTokenSilently } = useAuth0();
+  const gameId = useParams();
 
   const {
     register,
@@ -57,35 +38,33 @@ function EditCreatedGame() {
   }
 
   async function onSubmit(data, e) {
+    console.log("---");
     console.log(data);
+    console.log("---");
     setSubmitting(true);
     setPostError(null);
 
     const accessToken = await getAccessTokenSilently();
 
     try {
-      const response = await fetch(apiUrl, {
-        method: "POST",
-        headers: createHeaders(accessToken),
-        body: JSON.stringify({
-          id: 0,
-          gameTitle: data.gameTitle,
-          gameState: "REGISTRATION",
-          gameDescription: data.gameDescription,
-          nw_lat: data.nw_lat,
-          nw_lng: data.nw_lng,
-          se_lat: data.se_lat,
-          se_lng: data.se_lng,
-          players: [0],
-          missions: [0],
-        }),
-      });
+      const response = await fetch(
+        `https://hvz-api-noroff.herokuapp.com/api/v1/games/${gameId.gameId}`,
+        {
+          headers: createHeaders(accessToken),
+          method: "PUT",
+          body: JSON.stringify({
+            gameTitle: data.gameTitle,
+          }),
+        }
+      );
+      console.log(gameId);
+      console.log(data);
       setPostSuccess(true);
       e.target.reset();
       setTimeout(() => {
         setDisplayModalForm(false);
       }, 1500);
-      if (!response.ok) throw new Error("Could not create user with username");
+      if (response.ok) throw new Error("could not ");
       console.log(response);
       return [null, response];
     } catch (error) {
@@ -121,81 +100,6 @@ function EditCreatedGame() {
             />
             {errors.gameTitle && (
               <div className="mb-3 text-danger">{errors.gameTitle.message}</div>
-            )}
-            <Form.Label htmlFor="gameDescription" className="mt-3">
-              About the game
-            </Form.Label>
-            <Form.Control
-              {...register("gameDescription")}
-              id="gameDescription"
-              as="textarea"
-              rows={5}
-              placeholder="Describe the game - max 200 words"
-            />
-            {errors.gameDescription && (
-              <div className="mb-3 text-danger">
-                {errors.gameDescription.message}
-              </div>
-            )}
-            <Form.Label htmlFor="nw_lat" className="mt-3">
-              Northwest latitude
-            </Form.Label>
-            <Form.Control
-              {...register("nw_lat")}
-              id="nw_lat"
-              placeholder="nortwest latitude"
-            />
-            {errors.nw_lat && (
-              <div className="mb-3 text-danger">
-                {errors.nw_lat.message.includes("NaN")
-                  ? "Value must be a number (integer)"
-                  : errors.nw_lat.message}
-              </div>
-            )}
-            <Form.Label htmlFor="nw_lng" className="mt-3">
-              Northwest longitude
-            </Form.Label>
-            <Form.Control
-              {...register("nw_lng")}
-              id="nw_lng"
-              placeholder="nortwest longitude"
-            />
-            {errors.nw_lng && (
-              <div className="mb-3 text-danger">
-                {errors.nw_lng.message.includes("NaN")
-                  ? "Value must be a number (integer)"
-                  : errors.nw_lng.message}
-              </div>
-            )}
-            <Form.Label htmlFor="se_lat" className="mt-3">
-              Southeast latitude
-            </Form.Label>
-            <Form.Control
-              {...register("se_lat")}
-              id="se_lat"
-              placeholder="southeast latitude"
-            />
-            {errors.se_lat && (
-              <div className="mb-3 text-danger">
-                {errors.se_lat.message.includes("NaN")
-                  ? "Value must be a number (integer)"
-                  : errors.se_lat.message}
-              </div>
-            )}
-            <Form.Label htmlFor="se_lng" className="mt-3">
-              Southeast longitude
-            </Form.Label>
-            <Form.Control
-              {...register("se_lng")}
-              id="se_lng"
-              placeholder="southeast longitude"
-            />
-            {errors.se_lng && (
-              <div className="mb-3 text-danger">
-                {errors.se_lng.message.includes("NaN")
-                  ? "Value must be a number (integer)"
-                  : errors.se_lng.message}
-              </div>
             )}
 
             <button
