+++ src/components/admin/EditList.js
@@ -9,27 +9,10 @@ import { useNavigate, useParams } from "react-router-dom";
 import HeaderNavBar from "../nav/HeaderNavBar";
 import { useAuth0 } from "@auth0/auth0-react";
 import { createHeaders } from "../admin/CreateHeaders";
-import { Button } from "react-bootstrap";
-
 function EditList() {
-  const navigate = useNavigate();
   const { getAccessTokenSilently } = useAuth0();
   const gameId = useParams();
   const [gameData, setGame] = useState([]);
-
-  //Delete a game
-  let deleteGame = async () => {
-    const accessToken = await getAccessTokenSilently();
-    await fetch(
-      `${process.env.REACT_APP_API_SERVER_URL}games/${gameId.gameId}`,
-      {
-        headers: createHeaders(accessToken),
-        method: "DELETE",
-      }
-    );
-    navigate("/");
-    console.log("deleted");
-  };
 
   //Get game
   useEffect(() => {
@@ -56,9 +39,6 @@ function EditList() {
     <Accordion className="mt-2">
       <h1>test</h1>
       <HeaderNavBar title={gameData.gameTitle} />
-      <Button className="w-100 border-danger bg-danger" onClick={deleteGame}>
-        Delete current game
-      </Button>
       <Accordion.Item eventKey="0">
         <Accordion.Header className="text-end">Game Info</Accordion.Header>
         <Accordion.Body>
